refactor(luca-framework): extract CLI logic from index.ts to cli.ts
Move defineCommand/runMain/runInit logic to dedicated cli.ts file,
keeping index.ts as a pure barrel with only re-export statements.
